// frontend/src/app/api/models/termino-glosario-publico.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ContenidoRefPublico } from '../models/contenido-ref-publico';
import { Slug } from '../models/slug';
export interface TerminoGlosarioPublico {
  definicion: string;
  slug: Slug;
  termino: string;
  usado_en: Array<ContenidoRefPublico>;
}